refactor(directives): move to src
