--- packages/web/directives/index.ts
@@ -1,4 +1,0 @@
-export { default as ClickOutside } from './click-outside'
-export { default as RepeatClick } from './repeat-click'
-export { default as TrapFocus } from './trap-focus'
-export { default as Mousewheel } from './mousewheel'
